# PasswordManager.Backend/src/Password/PasswordManager.Password.Client/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PasswordManager.Password API",
    "version": "v1"
  },
  "paths": {
    "/api/password": {
      "post": {
        "tags": [
          "Password"
        ],
        "summary": "Create password",
        "description": "Creates a password",
        "operationId": "CreatePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePasswordRequestWithBody"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePasswordRequestWithBody"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePasswordRequestWithBody"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OperationAcceptedResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationAcceptedResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationAcceptedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/password/{passwordId}": {
      "delete": {
        "tags": [
          "Password"
        ],
        "summary": "Deletes a Password",
        "description": "Deletes a password",
        "operationId": "DeletePassword",
        "parameters": [
          {
            "name": "passwordId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "created-by-user-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OperationAcceptedResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationAcceptedResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationAcceptedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Password"
        ],
        "summary": "Get Password by Password id",
        "description": "Get Password by Password id",
        "operationId": "GetPassword",
        "parameters": [
          {
            "name": "passwordId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Password"
        ],
        "summary": "Update password",
        "description": "Update password details",
        "operationId": "UpdatePassword",
        "parameters": [
          {
            "name": "passwordId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "created-by-user-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePasswordRequestDetails"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePasswordRequestDetails"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePasswordRequestDetails"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OperationAcceptedResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationAcceptedResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationAcceptedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/password/generate": {
      "get": {
        "tags": [
          "Password"
        ],
        "summary": "Generates Password",
        "description": "Generates a password",
        "operationId": "GeneratePassword",
        "parameters": [
          {
            "name": "PasswordLength",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratePasswordResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratePasswordResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratePasswordResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/operations": {
      "get": {
        "tags": [
          "Operations"
        ],
        "summary": "Get operation by request id",
        "description": "Get operation by request id",
        "operationId": "GetOperationByRequestId",
        "parameters": [
          {
            "name": "requestId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/passwords/user": {
      "get": {
        "tags": [
          "Password"
        ],
        "summary": "Gets Passwords from UserId",
        "description": "Gets Passwords from UserId",
        "operationId": "GetPasswordsFromUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResponses"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResponses"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResponses"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/passwords/by-user-and-url": {
      "get": {
        "tags": [
          "Password"
        ],
        "summary": "Gets Passwords by UserId and Url",
        "description": "Retrieves passwords associated with a UserId that match a specified URL",
        "operationId": "GetPasswordsByUserIdAndUrl",
        "parameters": [
          {
            "name": "UserId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Url",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResponses"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResponses"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResponses"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreatePasswordRequestDetails": {
        "required": [
          "friendlyName",
          "password",
          "url",
          "userId",
          "username"
        ],
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "friendlyName": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "CreatePasswordRequestWithBody": {
        "required": [
          "createdByUserId"
        ],
        "type": "object",
        "properties": {
          "createdByUserId": {
            "minLength": 1,
            "type": "string"
          },
          "details": {
            "$ref": "#/components/schemas/CreatePasswordRequestDetails"
          }
        },
        "additionalProperties": false
      },
      "GeneratePasswordResponse": {
        "required": [
          "password"
        ],
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "OperationAcceptedResponse": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "OperationName": {
        "enum": [
          "CreatePassword",
          "UpdatePassword",
          "DeletePassword"
        ],
        "type": "string"
      },
      "OperationResponse": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string"
          },
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "operationName": {
            "$ref": "#/components/schemas/OperationName"
          },
          "operationStatus": {
            "$ref": "#/components/schemas/OperationStatus"
          },
          "createdBy": {
            "type": "string"
          },
          "createdUtc": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completedUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "data": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OperationStatus": {
        "enum": [
          "Queued",
          "Processing",
          "Completed",
          "Failed"
        ],
        "type": "string"
      },
      "PasswordResponse": {
        "required": [
          "friendlyName",
          "id",
          "password",
          "url",
          "userId",
          "username"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "url": {
            "type": "string"
          },
          "friendlyName": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "PasswordResponses": {
        "type": "object",
        "properties": {
          "passwordsResponses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PasswordResponse"
            }
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
      "UpdatePasswordRequestDetails": {
        "required": [
          "friendlyName",
          "password",
          "url",
          "username"
        ],
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "friendlyName": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "additionalProperties": false
      }
    }
  }
}